pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "brlok"
version = "0.1.0"
description = "Application d'entraînement bloc et pan - génération de séances"
readme = "README.md"
requires-python = ">=3.8"
license = { text = "MIT" }
dependencies = [
    "pyside6>=6.6",
    "typer>=0.9",
    "pydantic>=2.0",
    "platformdirs>=4.0",
    "odfpy>=1.4",
    "reportlab>=4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

[project.scripts]
brlok = "brlok.cli.commands:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["brlok*"]
